chessmates-server/requirements.txt
mysql-connector-python==8.0.29
Flask==2.1.2
chess==1.9.1
PyJWT==2.4.0
Flask-SocketIO==5.2.0
bcrypt==3.2.2
gunicorn==20.1.0
flask_cors==3.0.10
flask_session==0.4.0
gevent==21.12.0
gevent-websocket==0.10.1
flask_talisman==1.0.0